requirements-dev.txt
black==23.7.0
flake8==6.1.0
isort==5.12.0
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-playwright==0.4.2
